pyproject.toml
[tool.poetry]
name = "budgeting-server"
version = "0.0.1"
description = "FastAPI backend for the budgeting application for my thesis"
authors = ["Viktor Koch <koch.viktor007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
pydantic = {extras = ["email"], version = "^2.6.4"}
pydantic-settings = "^2.2.1"
arq = "^0.25.0"
fastapi = ">=0.100.0,<0.111.0"
fastcrud = "^0.11.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
